assign downtime reason updated

+++ app/Http/Controllers/StationOperatorController.php
@@ -133,7 +133,11 @@ class StationOperatorController extends Controller
         
         if ($stationOperator) {
             // remove previous operator
-            $stationOperator->delete();
+            if ($stationOperator->operator_id != $operatorId) {
+                $stationOperator->delete();
+            } else {
+                return TRUE;
+            }
         }
 
         $newStationOperator = new StationOperator();
